Split Jid struct into BareJid and FullJid. Jid is now an enum
This will help with being able to enforce the usage of bare or full at
compile time. It is still possible to allow one or the other with the
`Jid` enum.

Thanks to O01eg (from xmpp-rs@muc.linkmauve.fr) for the help. This
commit also contains code from them.

Signed-off-by: Maxime “pep” Buquet <pep@bouah.net>