minidom: Allow contextual NS to be passed when building element from reader
When reading from a (tcp|tls)stream, incoming stanzas generally look
like so: `<iq from='foo' type='result'>..</iq>` and don't have any
namespace declared because a default one is already declared on the
stream (generally `jabber:client` or similar).

It is currently not possible with minidom to get this element returned
with all the contextual namespaces we know, unless we pass the whole
tree (that is, including `<stream></stream>`).

This method might not be the best but allows us to build upon and can be
reworked later if necessary.

Signed-off-by: Maxime “pep” Buquet <pep@bouah.net>